requirements.txt
# Personal Intelligence System - Unified Requirements
# Python 3.8+

# Web Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
python-multipart>=0.0.6
starlette>=0.27.0

# NER & NLP
spacy>=3.7.0
jieba>=0.42.1
transformers>=4.35.0
torch>=2.1.0

# Embedding
sentence-transformers>=2.2.2

# LLM Services
openai>=1.3.0
anthropic>=0.8.0
ollama>=0.1.0

# Data & ML
numpy>=1.24.0
scikit-learn>=1.3.0
pandas>=2.0.0

# Utilities
pyyaml>=6.0.1
python-dotenv>=1.0.0
requests>=2.31.0
httpx>=0.25.0

# Privacy & Security
cryptography>=41.0.0

# Development
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0
